連続再生性能の向上
- 24h365d 連続運用を念頭に細部をチューニング
- 自動再起動(DDoS 配慮)
- リングバッファ拡張
- ESP32S3 + I2S を拡充
- 細かいバグの修正

+++ src/AudioFileSourceHLS.hpp
@@ -10,11 +10,11 @@
 
 class AudioFileSourceHLS : public AudioFileSource {
   public:
-    AudioFileSourceHLS(uint16_t _buffSize) : buffer(_buffSize) {
+    AudioFileSourceHLS(size_t _buffSize) : buffer(_buffSize) {
       init();
     }
     
-    AudioFileSourceHLS(uint8_t *_buffer, uint16_t _buffSize) : buffer(_buffer, _buffSize) {
+    AudioFileSourceHLS(uint8_t *_buffer, size_t _buffSize) : buffer(_buffer, _buffSize) {
       init();
     }
 
@@ -43,7 +43,7 @@ class AudioFileSourceHLS : public AudioFileSource {
       return src;
     }
 
-    virtual uint32_t getSize() override {
+    virtual size_t getSize() override {
       return buffer.available();
     }
 
@@ -51,11 +51,11 @@ class AudioFileSourceHLS : public AudioFileSource {
       return src && src->loop();
     }
 
-    virtual uint32_t read(void *dst, uint32_t length) override {
+    virtual size_t read(void *dst, size_t length) override {
       return readInternal(dst, length, false);
     }
 
-    virtual uint32_t readNonBlock(void *dst, uint32_t length) override {
+    virtual size_t readNonBlock(void *dst, size_t length) override {
       return readInternal(dst, length, true);
     }
 
@@ -114,7 +114,7 @@ class AudioFileSourceHLS : public AudioFileSource {
     SimpleRingBuffer<uint8_t> buffer;
     uint16_t timeout;
 
-    virtual uint32_t readInternal(void *dst, uint32_t length, bool nonblock) {
+    virtual size_t readInternal(void *dst, size_t length, bool nonblock) {
       auto rLen = buffer.read((uint8_t *)dst, length);
       if(length && rLen < length)
         cb.st(STATUS_TOO_SLOW, PSTR("receive is too slow.."));
