+++ src/AudioFileSourceTS.hpp
@@ -10,10 +10,10 @@
 
 class AudioFileSourceTS : public AudioFileSourceHLS {
   public:
-    AudioFileSourceTS(uint16_t _buffSize) : AudioFileSourceHLS(_buffSize) {}
-    AudioFileSourceTS(uint8_t *_buffer, uint16_t _buffSize) : AudioFileSourceHLS(_buffer, _buffSize) {}
-    AudioFileSourceTS(AudioFileSource * _src, uint16_t _buffSize) : AudioFileSourceHLS(_src, _buffSize) {}
-    AudioFileSourceTS(AudioFileSource * _src, uint8_t *_buffer, uint16_t _buffSize) : AudioFileSourceHLS(_src, _buffer, _buffSize) {}
+    AudioFileSourceTS(size_t _buffSize) : AudioFileSourceHLS(_buffSize) {}
+    AudioFileSourceTS(uint8_t *_buffer, size_t _buffSize) : AudioFileSourceHLS(_buffer, _buffSize) {}
+    AudioFileSourceTS(AudioFileSource * _src, size_t _buffSize) : AudioFileSourceHLS(_src, _buffSize) {}
+    AudioFileSourceTS(AudioFileSource * _src, uint8_t *_buffer, size_t _buffSize) : AudioFileSourceHLS(_src, _buffer, _buffSize) {}
 
     void identifyPid(uint16_t _pid) {
       PID = _pid;
